Quick fix in GaussData.py file related to the indexes

+++ utils/GaussData.py
@@ -14,9 +14,9 @@ class GaussData:
         self.dataset = pd.read_csv(str(file_path)).to_dict("records")
         self.dataset_length = len(self.dataset)
 
-        self.train_dataset = self.dataset[:(0.9 * self.dataset_length)]
-        self.val_dataset = self.dataset[(0.9 * self.dataset_length):(0.95 * self.dataset_length)]
-        self.test_dataset = self.dataset[(0.95 * self.dataset_length):]
+        self.train_dataset = self.dataset[:int(0.9 * self.dataset_length)]
+        self.val_dataset = self.dataset[int(0.9 * self.dataset_length):int(0.95 * self.dataset_length)]
+        self.test_dataset = self.dataset[int(0.95 * self.dataset_length):]
 
         self.train_dataloader = DataLoader(self.train_dataset, collate_fn=self.collate_fn, batch_size=BATCH_SIZE, shuffle=SHUFFLE, num_workers=NUM_WORKERS, pin_memory=True, drop_last=DROP_lAST)
         self.val_dataloader = DataLoader(self.val_dataset, collate_fn=self.collate_fn, batch_size=BATCH_SIZE, shuffle=SHUFFLE, num_workers=NUM_WORKERS, pin_memory=True, drop_last=DROP_lAST)
